# docs/server/swagger.json
{
    "components": {
        "schemas": {
...
            "github_com_stacklok_toolhive_pkg_client.RegisteredClient": {
                "properties": {
                    "groups": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "name": {
                        "$ref": "#/components/schemas/github_com_stacklok_toolhive_pkg_client.ClientApp"
                    }
                },
                "type": "object"
            },
...
            "pkg_api_v1.createClientRequest": {
                "properties": {
                    "groups": {
                        "description": "Groups is the list of groups configured on the client.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "name": {
                        "$ref": "#/components/schemas/github_com_stacklok_toolhive_pkg_client.ClientApp"
                    }
                },
                "type": "object"
            },
            "pkg_api_v1.createClientResponse": {
                "properties": {
                    "groups": {
                        "description": "Groups is the list of groups configured on the client.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "name": {
                        "$ref": "#/components/schemas/github_com_stacklok_toolhive_pkg_client.ClientApp"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "This is the ToolHive API server.",
        "title": "ToolHive API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/v1beta/clients": {
            "get": {
                "description": "List all registered clients in ToolHive",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/github_com_stacklok_toolhive_pkg_client.RegisteredClient"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List all clients",
                "tags": [
                    "clients"
                ]
            },
            "post": {
                "description": "Register a new client with ToolHive",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/pkg_api_v1.createClientRequest",
                                        "summary": "client",
                                        "description": "Client to register"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Client to register",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.createClientResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Invalid request or unsupported client type"
                    }
                },
                "summary": "Register a new client",
                "tags": [
                    "clients"
                ]
            }
        },
...
        "/api/v1beta/groups": {
            "get": {
                "description": "Get a list of all groups",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.groupListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List all groups",
                "tags": [
                    "groups"
                ]
            },
            "post": {
                "description": "Create a new group with the specified name",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/pkg_api_v1.createGroupRequest",
                                        "summary": "group",
                                        "description": "Group creation request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Group creation request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.createGroupResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Create a new group",
                "tags": [
                    "groups"
                ]
            }
        },
...
        "/api/v1beta/registry": {
            "get": {
                "description": "Get a list of the current registries",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.registryListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List registries",
                "tags": [
                    "registry"
                ]
            },
            "post": {
                "description": "Add a new registry",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "501": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Implemented"
                    }
                },
                "summary": "Add a registry",
                "tags": [
                    "registry"
                ]
            }
        },
...
        "/api/v1beta/registry/{name}/servers": {
            "get": {
                "description": "Get a list of servers in a specific registry",
                "parameters": [
                    {
                        "description": "Registry name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.listServersResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "List servers in a registry",
                "tags": [
                    "registry"
                ]
            }
        },
        "/api/v1beta/registry/{name}/servers/{serverName}": {
            "get": {
                "description": "Get details of a specific server in a registry",
                "parameters": [
                    {
                        "description": "Registry name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ImageMetadata name",
                        "in": "path",
                        "name": "serverName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.getServerResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Get a server from a registry",
                "tags": [
                    "registry"
                ]
            }
        },
        "/api/v1beta/secrets": {
            "post": {
                "description": "Setup the secrets provider with the specified type and configuration.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/pkg_api_v1.setupSecretsRequest",
                                        "summary": "request",
                                        "description": "Setup secrets provider request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Setup secrets provider request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/pkg_api_v1.setupSecretsResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Setup or reconfigure secrets provider",
                "tags": [
                    "secrets"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}